requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Django==5.2.9
djangorestframework==3.16.1
fastapi==0.123.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
pytest-django==4.11.1
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
sqlparse==0.5.4
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
